eval/requirements.txt
numpy==1.18.0
Pillow==7.0.0
PyYAML==5.2
tqdm==4.41.1
